requirements.txt
Flask
pandas
plotly
transformers
torch
python-dotenv
matplotlib
langchain
langchain_community
sentence-transformers
faiss-cpu
tensorflow-cpu
pydantic